backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# HTTP Clients
httpx==0.25.2
requests==2.31.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Geo/Maps
geopy==2.4.0
shapely==2.0.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
pytz==2023.3

# Environment
python-environ==0.4.54

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

